feat: integrated twilio api to text users magic links
Users are able to receive magic links via text via twilio. Additionally
the home page was adjusted to allow for users to re request links via text
in addition to just email. In addition, the RSVP page is now session, and email protected.
This means that only users with a valid email stored in the database are able to directly
access the form. If a valid email is not found - users are redirected to /complete-profile
such that they can enter their updated email , which is subsequently updated on both the guests and auth
tables.